backend/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
setuptools==68.2.2
Werkzeug==3.0.2
wheel==0.41.2
pandas
scipy
psycopg2-binary
python-dotenv
gunicorn